+++ Assets/Resources/Prefabs/Spawn.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 7a19586a13bf84453b7b438c5684dfb0
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
