+++ Assets/Scripts/Enemy.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5d93ad068826e4991bd192c978519e8a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
