+++ Assets/Scripts/Enemy/SpawnEnemy.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8428467b733184d5c99b15d064de3f54
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
