+++ UserSettings/EditorUserSettings.asset
@@ -12,13 +12,13 @@ EditorUserSettings:
       value: 065703530400085f5e0b0e7512710a454e164f797c29256128704b61b5b23639
       flags: 0
     RecentlyUsedSceneGuid-2:
-      value: 565752025453505a0e0b5b23427b0d48134e1b2e74297063752b4835e7b96c60
+      value: 5a5757560101590a5d0c0e24427b5d44434e4c7a7b7a23677f2b4565b7b5353a
       flags: 0
     RecentlyUsedSceneGuid-3:
-      value: 54030c505c0d5f0a09560d21477b0848174f402b7c7e25362b7c4c60b6b5366d
+      value: 565752025453505a0e0b5b23427b0d48134e1b2e74297063752b4835e7b96c60
       flags: 0
     RecentlyUsedSceneGuid-4:
-      value: 5a5757560101590a5d0c0e24427b5d44434e4c7a7b7a23677f2b4565b7b5353a
+      value: 54030c505c0d5f0a09560d21477b0848174f402b7c7e25362b7c4c60b6b5366d
       flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
